fedora.ks: custom partitioning
This ended up being slightly more complex than it probably needed. Once I added
a the second nvme for /home to lizzie, it dawned on me that I could always keep
/home and just reinstall the OS on the primary nvme as I see fit.

Then I thought what if I'd like to keep home on other systems aswell? And what
if they were bios based? And a whole other bunch of what ifs. And doing stuff
dynamically, and it got stupid fast.

lee is a r720xd with 12 drives in front, 1 of 2 possible in the rear, on a H310
mini mono flashed to IT mode. So the drives keep jumping around.

+++ .local/share/kickstart/fedora.ks
@@ -10,12 +10,6 @@ lang en_US.UTF-8
 # Run the Setup Agent on first boot
 firstboot --disable
 
-#ignoredisk --only-use=nvme0n1
-autopart
-
-# Partition clearing information
-clearpart --all --initlabel --drives=nvme0n1|sda|vda
-
 # System timezone
 timezone Europe/Oslo --utc
 
@@ -52,7 +46,7 @@ rpmfusion-nonfree-release
 
 %end
 
-%pre
+%pre --interpreter=/usr/bin/bash
 source /etc/os-release
 cat << EOF > /tmp/rpmfusion.repo
 repo --name=rpmfusion-free            --baseurl=http://download1.rpmfusion.org/free/fedora/releases/${VERSION_ID}/Everything/x86_64/os/
@@ -61,18 +55,78 @@ repo --name=rpmfusion-nonfree         --baseurl=http://download1.rpmfusion.org/n
 repo --name=rpmfusion-nonfree-updates --baseurl=http://download1.rpmfusion.org/nonfree/fedora/updates/${VERSION_ID}/x86_64/
 EOF
 
+
+declare -A uuid_host=(
+  [4c4c4544-0051-3810-8036-b9c04f5a5831]="lee"
+  [30be1077-acee-47dc-ba5f-785287a8684c]="lizzie"
+)
+
+declare -A host_disks=(
+  [lee]="/dev/disk/by-path/00.0-sas-exp0x500056b36789abff-phy10-lun-0"
+  [lizzie]="/dev/disk/by-path/pci-0000:00:0e.0-pci-10000:e1:00.0-nvme-1 /dev/disk/by-path/pci-0000:00:0e.0-pci-10000:e2:00.0-nvme-1 /dev/disk/by-path/pci-0000:00:0e.0-pci-10000:e2:00.0-nvme-1-part1"
+)
+
+host="${uuid_host[$(dmidecode -s system-uuid)]:-unknown}"
+disks=(${host_disks[$host]})
+disk_count=$(( ${#disks[@]} - 1))
+clearpart_count=$disk_count
+
+if grep -qw inst.keephome /proc/cmdline; then
+  keephome=1
+  ((clearpart_count -= 1)) # If we're keeping home, we keed the partition layout on that disk
+fi
+
+echo "bootloader --append=\"mitigations=off\"" > /tmp/partitioning.ks
+case "$host" in
+  lee)
+    cat << EOF >> /tmp/partitioning.ks
+ignoredisk --only-use=${disks[0]}
+clearpart --drives=${disks[0]} --all --initlabel
+autopart
+EOF
+  lizzie)
+    (IFS=','; echo "ignoredisk --only-use=${disks[*]:0:${disk_count}}" >> /tmp/partitioning.ks)
+    (IFS=','; echo "clearpart --drives=${disks[*]:0:${clearpart_count}} --all --initlabel" >> /tmp/partitioning.ks)
+
+    if [ -d /sys/firmware/efi ]; then
+      echo "part /boot/efi --ondisk=${disks[0]} --fstype=\"efi\" --size=600 --label=uefi--fsoptions=\"umask=0077,shortname=winnt\"" >> /tmp/partitioning.ks
+    else
+      echo "part biosboot --ondisk=${disks[0]} --fstype=\"biosboot\" --size=1" >> /tmp/partitioning.ks
+    fi
+
+    cat << EOF >> /tmp/partitioning.ks
+part /boot --ondisk=${disks[0]} --fstype="ext4" --size=1024 --label=boot --fsoptions="defaults,discard"
+part /     --ondisk=${disks[0]} --fstype="xfs"  --grow      --label=root --fsoptions="defaults,discard,noatime"
+EOF
+    if (( keephome )); then
+      echo "part /home --onpart=${disks[2]} --fstype=\"xfs\" --grow --label=home --fsoptions=\"defaults,discard,noatime\" --noformat >> /tmp/partitioning.ks
+    else
+      echo "part /home --ondisk=${disks[1]} --fstype=\"xfs\" --grow --label=home --fsoptions=\"defaults,discard,noatime\"" >> /tmp/partitioning.ks
+    fi
+    ;;
+  *)
+    cat << EOF > /tmp/partitioning.ks
+autopart
+clearpart --drives=nvme0n1|sda|vda --all --initlabel
+EOF
+    ;;
+esac
+
 %end
 
 %include /tmp/rpmfusion.repo
+%include /tmp/partitioning.ks
 
 %post --interpreter=/usr/bin/bash --log=/root/ks-post.log
 chage -d 0 erikberg
-su - erikberg -c 'git clone --bare https://github.com/eb4x/dotfiles.git $HOME/.dotfiles'
-su - erikberg -c 'rm $HOME/.bashrc $HOME/.bash_profile'
-su - erikberg -c 'git --git-dir=$HOME/.dotfiles --work-tree=$HOME checkout'
-su - erikberg -c 'git --git-dir=$HOME/.dotfiles --work-tree=$HOME config --local status.showUntrackedFiles no'
-su - erikberg -c 'git --git-dir=$HOME/.dotfiles --work-tree=$HOME config --local user.email github@slipsprogrammor.no'
-su - erikberg -c 'restorecon -R $HOME'
+if ! grep -qw inst.keephome /proc/cmdline; then
+  su - erikberg -c 'git clone --bare https://github.com/eb4x/dotfiles.git $HOME/.dotfiles'
+  su - erikberg -c 'rm $HOME/.bashrc $HOME/.bash_profile'
+  su - erikberg -c 'git --git-dir=$HOME/.dotfiles --work-tree=$HOME checkout'
+  su - erikberg -c 'git --git-dir=$HOME/.dotfiles --work-tree=$HOME config --local status.showUntrackedFiles no'
+  su - erikberg -c 'git --git-dir=$HOME/.dotfiles --work-tree=$HOME config --local user.email github@slipsprogrammor.no'
+  su - erikberg -c 'restorecon -R $HOME'
+fi
 %end
 
 # Completion methods
